+++ src/main/java/edu/icet/service/customer/AnniversaryEventService.java
@@ -1,14 +1,13 @@
 package edu.icet.service.customer;
 
 import edu.icet.dto.Anniversary;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 public interface AnniversaryEventService {
     boolean add (Anniversary anniversary);
     List<Anniversary> getAll();
-    boolean delete(Integer eventId);
+    boolean delete(Long eventId);
     boolean update(Anniversary anniversary);
-    Anniversary get(Integer eventId);
+    Anniversary get(Long eventId);
 }
